Added Dijkstra algorithm code in cpp.
I think this code is not efficient enough because taking the new node with the lowest value requires to iterate over all  unvisited nodes. I think one solution is to override unvisited_node set's comparator so that it compares node values but when I implement it some errors occur and I can't fix them.