Added Merge Sort algorithm code in cpp.
Getting the sub arrays of the main array is a bit not optimal (because it allocates two new arrays) and may be solved by making merge sort to take first and last index as an argument so that it only sorts array in that margins. That implementation would make code more efficient in time and also in memory.